Add function to get fitted values as a data frame
This commit adds the `dfa_fitted` function, which returns the fitted values from
a DFA as a data frame. Previously, a data frame of the fitted values was created
in the `plot_fitted` function, but it wasn't possible to return the data frame.
The `plot_fitted` function was also updated to use the output of `dfa_fitted` to
generate the plotting data.

The `dfa_fitted` function includes a `conf_level` argument, allowing the user to
specify the CI level to compute. The `conf_level` argument was also added to the
`plot_fitted` function.