Add function to get trends as a data frame
This commit adds the `dfa_trends` function, which returns the trends from a DFA
as a data frame. Previously, a data frame of the trends was created in
the `plot_trends` function, but it wasn't possible to return the data frame. The
`plot_trends` function was also updated to use the output of `dfa_trends` to
generate the plotting data.