Add function to get loadings as a data frame
This commit adds the `dfa_loadings` function, which returns the loadings from a
DFA as a data frame. Previously, a data frame of the loadings was created in the
`plot_loadings` function, but it wasn't possible to return the data frame. The
`plot_loadings` function was also updated to use the output of `dfa_loadings` to
generate the plotting data.